refactor(commands): extract setup/review commands into setup.rs
