+++ src-rust/crates/commands/src/lib.rs
@@ -252,6 +252,8 @@ mod display;
 pub use display::*;
 mod maintenance;
 pub use maintenance::*;
+mod setup;
+pub use setup::*;
 mod ui_settings;
 use ui_settings::*;
 
@@ -277,9 +279,6 @@ pub struct ImportConfigCommand;
 pub struct ThinkingCommand;
 // New commands
 // Batch-1 new commands
-pub struct StatuslineCommand;
-pub struct SecurityReviewCommand;
-pub struct TerminalSetupCommand;
 pub struct ExtraUsageCommand;
 pub struct FastCommand;
 pub struct FeedbackCommand;
@@ -1200,245 +1199,6 @@ impl SlashCommand for ThinkingCommand {
                 model
             ))
         }
-    }
-}
-
-// ---- /statusline ---------------------------------------------------------
-
-#[async_trait]
-impl SlashCommand for StatuslineCommand {
-    fn name(&self) -> &str { "statusline" }
-    fn description(&self) -> &str { "Configure what is shown in the status line" }
-    fn help(&self) -> &str {
-        "Usage: /statusline [show|hide] [cost|tokens|model|time|all]\n\n\
-         Controls which items appear in the TUI status bar at the bottom.\n\
-         Settings are persisted to ~/.claurst/ui-settings.json.\n\n\
-         Examples:\n\
-           /statusline               — show current configuration\n\
-           /statusline show cost     — show cost in status line\n\
-           /statusline hide tokens   — hide token count\n\
-           /statusline show all      — show everything\n\
-           /statusline hide all      — hide everything"
-    }
-
-    async fn execute(&self, args: &str, _ctx: &mut CommandContext) -> CommandResult {
-        let args = args.trim();
-        let current = load_ui_settings();
-
-        if args.is_empty() {
-            return CommandResult::Message(format!(
-                "Status line configuration\n\
-                 ─────────────────────────\n\
-                 Show cost:   {cost}\n\
-                 Show tokens: {tokens}\n\
-                 Show model:  {model}\n\
-                 Show time:   {time}\n\n\
-                 Use /statusline [show|hide] [cost|tokens|model|time|all] to change.",
-                cost = fmt_bool(current.statusline_show_cost.unwrap_or(true)),
-                tokens = fmt_bool(current.statusline_show_tokens.unwrap_or(true)),
-                model = fmt_bool(current.statusline_show_model.unwrap_or(true)),
-                time = fmt_bool(current.statusline_show_time.unwrap_or(true)),
-            ));
-        }
-
-        let mut parts = args.splitn(2, ' ');
-        let verb = parts.next().unwrap_or("").trim();
-        let item = parts.next().unwrap_or("").trim();
-
-        let show = match verb {
-            "show" | "enable" | "on" => true,
-            "hide" | "disable" | "off" => false,
-            _ => {
-                return CommandResult::Error(
-                    "Usage: /statusline [show|hide] [cost|tokens|model|time|all]".to_string(),
-                )
-            }
-        };
-
-        if item.is_empty() || item == "all" {
-            match mutate_ui_settings(|s| {
-                s.statusline_show_cost = Some(show);
-                s.statusline_show_tokens = Some(show);
-                s.statusline_show_model = Some(show);
-                s.statusline_show_time = Some(show);
-            }) {
-                Ok(_) => return CommandResult::Message(format!(
-                    "Status line: all items {}.",
-                    if show { "shown" } else { "hidden" }
-                )),
-                Err(e) => return CommandResult::Error(format!("Failed to save: {}", e)),
-            }
-        }
-
-        let result = match item {
-            "cost" => mutate_ui_settings(|s| s.statusline_show_cost = Some(show)),
-            "tokens" | "token" => mutate_ui_settings(|s| s.statusline_show_tokens = Some(show)),
-            "model" => mutate_ui_settings(|s| s.statusline_show_model = Some(show)),
-            "time" | "clock" => mutate_ui_settings(|s| s.statusline_show_time = Some(show)),
-            other => {
-                return CommandResult::Error(format!(
-                    "Unknown item '{}'. Use: cost, tokens, model, time, or all.",
-                    other
-                ))
-            }
-        };
-
-        match result {
-            Ok(_) => CommandResult::Message(format!(
-                "Status line: {} {}.",
-                item,
-                if show { "shown" } else { "hidden" }
-            )),
-            Err(e) => CommandResult::Error(format!("Failed to save: {}", e)),
-        }
-    }
-}
-
-fn fmt_bool(v: bool) -> &'static str {
-    if v { "on" } else { "off" }
-}
-
-// ---- /security-review ----------------------------------------------------
-
-#[async_trait]
-impl SlashCommand for SecurityReviewCommand {
-    fn name(&self) -> &str { "security-review" }
-    fn description(&self) -> &str { "Run a security review of the current project" }
-    fn help(&self) -> &str {
-        "Usage: /security-review [path]\n\n\
-         Asks Claurst to perform a security review of the codebase.\n\
-         Analyzes for common vulnerabilities: injection attacks, auth issues,\n\
-         secrets exposure, unsafe deserialization, path traversal, etc."
-    }
-
-    async fn execute(&self, args: &str, ctx: &mut CommandContext) -> CommandResult {
-        let target = if args.trim().is_empty() {
-            ctx.working_dir.display().to_string()
-        } else {
-            args.trim().to_string()
-        };
-
-        CommandResult::UserMessage(format!(
-            "Please perform a comprehensive security review of the code in `{target}`.\n\n\
-             Focus on identifying:\n\
-             1. Injection vulnerabilities (SQL, command, LDAP, XSS, SSTI)\n\
-             2. Authentication and authorization flaws\n\
-             3. Hardcoded secrets, API keys, or passwords\n\
-             4. Insecure deserialization\n\
-             5. Path traversal or file inclusion vulnerabilities\n\
-             6. Cryptographic weaknesses (weak algorithms, bad IV usage, key reuse)\n\
-             7. Dependency vulnerabilities (check for outdated packages)\n\
-             8. Race conditions and TOCTOU issues\n\
-             9. Information disclosure (verbose errors, debug endpoints)\n\
-             10. Any OWASP Top 10 issues relevant to this codebase\n\n\
-             For each finding, provide:\n\
-             - Severity: Critical/High/Medium/Low/Informational\n\
-             - File and line number\n\
-             - Description of the vulnerability\n\
-             - Proof of concept or reproduction steps\n\
-             - Recommended remediation\n\n\
-             Start by reading the main source files and any dependency manifests.",
-            target = target,
-        ))
-    }
-}
-
-// ---- /terminal-setup -----------------------------------------------------
-
-#[async_trait]
-impl SlashCommand for TerminalSetupCommand {
-    fn name(&self) -> &str { "terminal-setup" }
-    fn description(&self) -> &str { "Help configure your terminal for optimal Claurst use" }
-    fn help(&self) -> &str {
-        "Usage: /terminal-setup\n\n\
-         Diagnoses your terminal environment and gives recommendations for\n\
-         optimal Claurst display (font, color support, Unicode, etc.)."
-    }
-
-    async fn execute(&self, _args: &str, _ctx: &mut CommandContext) -> CommandResult {
-        let mut checks: Vec<String> = Vec::new();
-
-        // Check TERM variable
-        let term = std::env::var("TERM").unwrap_or_default();
-        let colorterm = std::env::var("COLORTERM").unwrap_or_default();
-        let term_program = std::env::var("TERM_PROGRAM").unwrap_or_default();
-
-        // Terminal identification
-        let terminal_name = if !term_program.is_empty() {
-            term_program.clone()
-        } else {
-            term.clone()
-        };
-        checks.push(format!("Terminal:      {}", terminal_name));
-
-        // Color depth
-        let color_depth = if colorterm == "truecolor" || colorterm == "24bit" {
-            "24-bit true color (optimal)"
-        } else if term.contains("256color") || colorterm == "256color" {
-            "256 colors (good)"
-        } else if !term.is_empty() {
-            "Basic colors (limited)"
-        } else {
-            "Unknown"
-        };
-        checks.push(format!("Colors:        {}", color_depth));
-
-        // Check if UNICODE is likely supported
-        let lang = std::env::var("LANG").unwrap_or_default();
-        let lc_all = std::env::var("LC_ALL").unwrap_or_default();
-        let unicode_env = lang.to_lowercase().contains("utf") || lc_all.to_lowercase().contains("utf");
-        checks.push(format!(
-            "Unicode/UTF-8: {}",
-            if unicode_env { "likely supported (LANG/LC_ALL contains UTF)" } else { "check LANG env var" }
-        ));
-
-        // Check for known good terminals
-        let is_good_terminal = matches!(
-            term_program.to_lowercase().as_str(),
-            "iterm.app" | "iterm2" | "hyper" | "warp" | "alacritty" | "kitty" | "wezterm"
-        ) || term_program.to_lowercase().contains("vscode")
-          || term_program.to_lowercase().contains("terminal");
-
-        checks.push(format!(
-            "Terminal type: {}",
-            if is_good_terminal { "well-known terminal (good)" } else { "verify settings below" }
-        ));
-
-        // Shell detection
-        let shell = std::env::var("SHELL").unwrap_or_else(|_| "unknown".to_string());
-        checks.push(format!("Shell:         {}", shell));
-
-        // Check for Nerd Fonts (heuristic: environment variable set by some terminals)
-        let nerd_font = std::env::var("NERD_FONT").is_ok()
-            || std::env::var("TERM_NERD_FONT").is_ok();
-
-        CommandResult::Message(format!(
-            "Terminal Setup Diagnostic\n\
-             ─────────────────────────\n\
-             {checks}\n\n\
-             Recommendations for optimal Claurst experience:\n\
-             ─────────────────────────────────────────────────\n\
-             1. Font: Use a Nerd Font for box-drawing characters and icons\n\
-                {nerd_hint}\n\
-                Download: https://www.nerdfonts.com/\n\
-             2. Color: Enable 24-bit true color:\n\
-                export COLORTERM=truecolor\n\
-             3. Unicode: Ensure UTF-8 locale:\n\
-                export LANG=en_US.UTF-8\n\
-             4. Recommended terminals:\n\
-                - WezTerm (all platforms)\n\
-                - Alacritty (all platforms)\n\
-                - Kitty (macOS/Linux)\n\
-                - Windows Terminal (Windows)\n\
-                - iTerm2 (macOS)\n\
-             5. Set terminal to unlimited scrollback for long conversations",
-            checks = checks.join("\n  "),
-            nerd_hint = if nerd_font {
-                "[ok] Nerd Font detected"
-            } else {
-                "[!] Nerd Font not detected — box-drawing may appear broken"
-            },
-        ))
     }
 }
